split request/response files
